monitor/requirements.txt
PyYAML==6.0.1
pysqlite3==0.5.2
